Convert CA1033's analyzer to an IOperation analyzer and move it to the APIDesignGuidelines analyzer.
