--- src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/CSharpInterfaceMethodsShouldBeCallableByChildTypes.cs
@@ -1,19 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System.Diagnostics;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Diagnostics;
-
-namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                          
-    /// <summary>
-    /// CA1033: Interface methods should be callable by child types
-    /// </summary>
-    [DiagnosticAnalyzer(LanguageNames.CSharp)]
-    public sealed class CSharpInterfaceMethodsShouldBeCallableByChildTypesAnalyzer : InterfaceMethodsShouldBeCallableByChildTypesAnalyzer
-    {
-
-    }
-}
