+++ src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.vbproj
@@ -118,7 +118,6 @@
     <Compile Include="BasicMarkAssembliesWithAssemblyVersion.Fixer.vb" />
     <Compile Include="BasicDefineAccessorsForAttributeArguments.vb" />
     <Compile Include="BasicUsePropertiesWhereAppropriate.Fixer.vb" />
-    <Compile Include="BasicInterfaceMethodsShouldBeCallableByChildTypes.vb" />
     <Compile Include="BasicInterfaceMethodsShouldBeCallableByChildTypes.Fixer.vb" />
     <Compile Include="BasicOverrideMethodsOnComparableTypes.Fixer.vb" />
     <Compile Include="BasicStaticHolderTypesShouldBeSealed.vb" />
